requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
supabase==2.0.2
python-gotrue==1.4.0
librosa==0.10.0
numpy==1.24.0
openai-whisper>=20240930
scipy==1.10.0
pronouncing==0.2.0
python-dotenv==1.0.0
Werkzeug==2.3.0
setuptools>=65.0
requests>=2.31.0
